exec httplb process

+++ shell/main.go
@@ -36,6 +36,7 @@ import (
 	"github.com/ossrs/go-oryx/kernel"
 	"os"
 	"os/exec"
+	"sync"
 )
 
 var signature = fmt.Sprintf("SHELL/%v", kernel.Version())
@@ -50,6 +51,11 @@ type ShellConfig struct {
 		// The binary to exec.
 		binary string
 	} `json:"rtmplb"`
+	Httplb struct {
+		Enabled bool   `json:"enabled"`
+		Binary  string `json:"binary"`
+		Config  string `json:"config"`
+	} `json:"httplb"`
 }
 
 func (v *ShellConfig) String() string {
@@ -89,44 +95,87 @@ func (v *ShellConfig) Loads(c string) (err error) {
 			return
 		}
 	}
+	if r := &v.Httplb; r.Enabled {
+		if len(r.Binary) == 0 {
+			return fmt.Errorf("Empty httplb binary")
+		}
+		if r.Binary, err = exec.LookPath(r.Binary); err != nil {
+			ol.E(nil, fmt.Sprintf("Invalid httplb binary=%v, err is %v", r.Binary, err))
+			return
+		}
+		if _, err = os.Lstat(r.Config); err != nil {
+			ol.E(nil, fmt.Sprintf("Invalid httplb config=%v, err is %v", r.Config, err))
+			return
+		}
+	}
 
 	return
+}
+
+// The process type.
+type ProcessType int
+
+const (
+	ProcessRtmpLb ProcessType = 100 + iota
+	ProcessHttpLb
+)
+
+func (v ProcessType) String() string {
+	switch v {
+	case ProcessRtmpLb:
+		return "rtmplb"
+	case ProcessHttpLb:
+		return "httplb"
+	default:
+		return "other"
+	}
 }
 
 // The shell to exec all processes.
 type ShellBoss struct {
 	conf   *ShellConfig
 	rtmplb *exec.Cmd
+	httplb *exec.Cmd
 	ctx    ol.Context
+	q      chan ProcessType
+	wait   *sync.WaitGroup
 }
 
 func NewShellBoss(conf *ShellConfig) *ShellBoss {
 	return &ShellBoss{
 		conf: conf,
 		ctx:  &kernel.Context{},
+		q:    make(chan ProcessType, 2),
+		wait: &sync.WaitGroup{},
 	}
 }
 
 func (v *ShellBoss) Close() (err error) {
 	ctx := v.ctx
 
-	if v.conf.Rtmplb.Enabled {
-		if r := v.rtmplb.ProcessState; r == nil || !r.Exited() {
-			var r0, r1 error
+	// only kill process, for we use isolate goroutine to wait them.
+	if r := v.rtmplb.ProcessState; v.conf.Rtmplb.Enabled && (r == nil || !r.Exited()) {
+		var r0 error
 
-			if r0 = v.rtmplb.Process.Kill(); err == nil {
-				err = r0
-			}
-
-			if _, r1 = v.rtmplb.Process.Wait(); err == nil {
-				err = r1
-			}
-
-			ol.W(ctx, fmt.Sprintf("Shell: rtmplb exited, r0=%v, r1=%v, err is %v", r0, r1, err))
-		} else {
-			ol.T(ctx, "Shell: rtmplb already exited")
+		if r0 = v.rtmplb.Process.Kill(); err == nil {
+			err = r0
 		}
+
+		ol.W(ctx, fmt.Sprintf("Shell: rtmplb killed, r0=%v, err is %v", r0, err))
 	}
+
+	if r := v.httplb.ProcessState; v.conf.Httplb.Enabled && (r == nil || !r.Exited()) {
+		var r0 error
+
+		if r0 = v.httplb.Process.Kill(); err == nil {
+			err = r0
+		}
+
+		ol.W(ctx, fmt.Sprintf("Shell: httplb killed, r0=%v, err is %v", r0, err))
+	}
+
+	// wait for all process to quit.
+	v.wait.Wait()
 
 	return
 }
@@ -140,29 +189,77 @@ func (v *ShellBoss) ExecBuddies() (err error) {
 			ol.E(ctx, "Shell: exec rtmplb failed, err is", err)
 			return
 		}
+		v.wait.Add(1)
 
 		p := v.rtmplb
 		ol.T(ctx, fmt.Sprintf("Shell: exec rtmplb ok, args=%v, pid=%v", p.Args, p.Process.Pid))
 	}
+
+	if r := &v.conf.Httplb; r.Enabled {
+		v.httplb = exec.Command(r.Binary, "-c", r.Config)
+		if err = v.httplb.Start(); err != nil {
+			ol.E(ctx, "Shell: exec httplb failed, err is", err)
+			return
+		}
+		v.wait.Add(1)
+
+		p := v.httplb
+		ol.T(ctx, fmt.Sprintf("Shell: exec httplb ok, args=%v, pid=%v", p.Args, p.Process.Pid))
+	}
+
+	go func() {
+		defer v.wait.Done()
+		defer func() {
+			v.q <- ProcessRtmpLb
+		}()
+
+		ctx := v.ctx
+
+		if !v.conf.Rtmplb.Enabled {
+			return
+		}
+
+		if err = v.rtmplb.Wait(); err != nil {
+			ol.E(ctx, "Shell: rtmplb failed, err is", err)
+			return
+		}
+
+		err = fmt.Errorf("rtmplb exited")
+		ol.E(ctx, fmt.Sprintf("Shell: rtmplb exited, err is", err))
+	}()
+
+	go func() {
+		defer v.wait.Done()
+		defer func() {
+			v.q <- ProcessHttpLb
+		}()
+
+		ctx := v.ctx
+
+		if !v.conf.Httplb.Enabled {
+			return
+		}
+
+		if err = v.httplb.Wait(); err != nil {
+			ol.E(ctx, "Shell: httplb failed, err is", err)
+			return
+		}
+
+		err = fmt.Errorf("httplb exited")
+		ol.E(ctx, fmt.Sprintf("Shell: httplb exited, err is", err))
+	}()
 	return
 }
 
-func (v *ShellBoss) WaitRtmplb() (err error) {
+func (v *ShellBoss) Wait() {
 	ctx := v.ctx
 
-	if !v.conf.Rtmplb.Enabled {
+	if v, ok := <-v.q; !ok {
+		return
+	} else if v == ProcessRtmpLb || v == ProcessHttpLb {
+		ol.W(ctx, "kernel process", v, "quit, shell quit.")
 		return
 	}
-
-	if err = v.rtmplb.Wait(); err != nil {
-		ol.E(ctx, "Shell: rtmplb failed, err is", err)
-		return
-	}
-
-	err = fmt.Errorf("rtmplb exited")
-	ol.E(ctx, fmt.Sprintf("Shell: rtmplb exited, err is", err))
-
-	return
 }
 
 func main() {
@@ -187,10 +284,7 @@ func main() {
 	}
 	defer shell.Close()
 
-	if err = shell.WaitRtmplb(); err != nil {
-		ol.E(ctx, "Shell: rtmplb failed, err is", err)
-		return
-	}
+	shell.Wait()
 
 	return
 }
